pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speek-input"
version = "1.0.0"
description = "Linux voice input method with hotkey activation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "PyQt5>=5.15",
    "pynput>=1.7",
    "pyaudio>=0.2.13",
    "numpy>=1.20",
    "pyyaml>=6.0",
    "openai>=1.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "pytest>=7.0",
]
faster-whisper = [
    "faster-whisper>=0.10",
]

[project.scripts]
speek-input = "speek_input.main:main"

[project.gui-scripts]
speek-input-gui = "speek_input.main:main"

[tool.setuptools.packages.find]
where = ["src"]
